user/requirements.txt
fastapi[standard]
pika
psycopg2-binary
